+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Notable changes to this project are documented in this file.
 
 # Unreleased
+**Important note**: starting with this version, there is no more custom Markdown plugin required to support Markdown rendering. If you used projector-installer before this version, please manually remove custom Projector Markdown plugin from the IDE and enable bundled Markdown plugin.
+
 ## Added 
  - add fqdn of host to SAN certificates
  - output Projector server stdout to console on error